+++ R/syntopic_table_functions.R
@@ -8,7 +8,7 @@
 #' @param group_cols A vector of columns to group the surveyData by and from which to compose syntopic tables
 #' @param species_col_name The name of the species column
 #' @param plot_col_name The name of the plot ID column
-#' @param numeric_constancy If TRUE the ordinal constancy classes (I, II, III, IV, V) are transformed into numeric values (1, 2, 3, 4, 5).
+#' @param numeral_constancy If TRUE the numeric constancy classes (1, 2, 3, 4, 5) are transformed into roman numeral values (I, II, III, IV, V).
 #'
 #' @return A three column data frame containing the groups ID, species, and constancy.
 #' @export
@@ -18,8 +18,8 @@
 #'                               group_cols = c("Year", "Group"), 
 #'                               species_col_name = "Species", 
 #'                               plot_col_name = "Quadrat",
-#'                               numeric_constancy = FALSE)
-composeSyntopicTables <- function(surveyData, group_cols, species_col_name = "Species", plot_col_name = "Quadrat", numeric_constancy = FALSE){
+#'                               numeral_constancy = FALSE)
+composeSyntopicTables <- function(surveyData, group_cols, species_col_name = "Species", plot_col_name = "Quadrat", numeral_constancy = FALSE){
   
   # Determine the total number of quadrats per group
   plot_n <- surveyData |>
@@ -44,32 +44,33 @@ composeSyntopicTables <- function(surveyData, group_cols, species_col_name = "Sp
     dplyr::mutate(
       "Constancy" =
         dplyr::case_when(
-          `Relative Frequency` <= 0.2 ~ "I",
-          `Relative Frequency` <= 0.4 ~ "II",
-          `Relative Frequency` <= 0.6 ~ "III",
-          `Relative Frequency` <= 0.8 ~ "IV",
-          `Relative Frequency` <= 1.0 ~ "V",
-          TRUE ~ as.character(`Relative Frequency`)
+          `Relative Frequency` <= 0.2 ~ 1,
+          `Relative Frequency` <= 0.4 ~ 2,
+          `Relative Frequency` <= 0.6 ~ 3,
+          `Relative Frequency` <= 0.8 ~ 4,
+          `Relative Frequency` <= 1.0 ~ 5,
+          TRUE ~ as.numeric(`Relative Frequency`)
         )
     ) |>
     dplyr::select(ID, species_col_name, Constancy) |>
-    dplyr::mutate("Constancy" = factor(Constancy, levels = c("V", "IV", "III", "II", "I"))) |>
+    dplyr::mutate("Constancy" = factor(Constancy, levels = c(5, 4, 3, 2, 1))) |>
     dplyr::arrange(ID, Constancy, species_col_name)
   
-  if(numeric_constancy == TRUE){
+  if(numeral_constancy == TRUE){
     
     syntopicTables <- syntopicTables |>
       dplyr::mutate(
         "Constancy" = 
           dplyr::case_when(
-            Constancy == "I" ~ 1,
-            Constancy == "II" ~ 2,
-            Constancy == "III" ~ 3,
-            Constancy == "IV" ~ 4,
-            Constancy == "V" ~ 5,
-            TRUE ~ as.numeric(0)
+            Constancy == 1 ~ "I",
+            Constancy == 2 ~ "II",
+            Constancy == 3 ~ "III",
+            Constancy == 4 ~ "IV",
+            Constancy == 5 ~ "V",
+            TRUE ~ NA
           )
-      )
+      ) |>
+      dplyr::mutate("Constancy" = factor(Constancy, levels = c("V", "IV", "III", "II", "I")))
     
   }