+++ tests/testthat/test-syntopic_table_functions.R
@@ -4,7 +4,7 @@ testthat::test_that("composeSyntopicTables works", {
                                           group_cols = c("Year", "Group"), 
                                           species_col_name = "Species", 
                                           plot_col_name = "Quadrat",
-                                          numeric_constancy = FALSE)
+                                          numeral_constancy = FALSE)
   
   actual_colnames <- colnames(actual)
   actual_coltypes <- sapply(actual, class) |> unname()
@@ -15,7 +15,7 @@ testthat::test_that("composeSyntopicTables works", {
   testthat::expect_equal(actual_colnames, expected_colnames)
   testthat::expect_equal(actual_coltypes, expected_coltypes)
   
-  testthat::expect_true(all(actual$Constancy %in% c("I", "II", "III", "IV", "V")))
+  testthat::expect_true(all(actual$Constancy %in% c(1, 2, 3, 4, 5)))
   
   
 })